+++ examples/testmodule/tests/test_dryrun.py
@@ -4,8 +4,6 @@
     License: Apache 2.0
 """
 
-from inmanta.data.model import AttributeStateChange
-
 
 def test_dryrun(project):
     basemodel = """
@@ -19,9 +17,4 @@ def test_dryrun(project):
     changes = project.dryrun_resource("testmodule::Resource")
     assert ["value"] == list(changes.keys())
     change = changes["value"]
-    # change in type in iso7
-    if isinstance(change, AttributeStateChange):
-        # dict method is deprecated on pydanticv2, iso7,
-        # but replacement doesn't exist on pydanticv1, iso6
-        change = change.dict()
-    assert change == {"current": "read", "desired": "write"}
+    assert change.model_dump() == {"current": "read", "desired": "write"}